# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 知行创境教育创作平台 API
  version: 1.0.0
  description: |
    基于《知行创境》教育创作平台 V2.0 PRD 整理的后端接口 OpenAPI 文档。
    基础路径为 `/api/v1`，认证方式为 Bearer Token。
servers:
  - url: /api/v1
    description: Default API Server
...
paths:
...
  /exhibitions:
    get:
      tags: [Exhibitions]
      summary: 获取我的展厅列表
      operationId: getMyExhibitions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: taskId
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ExhibitionStatus'
        - in: query
          name: keyword
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 展厅分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedExhibitionSummary'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Exhibitions]
      summary: 创建展厅
      operationId: createExhibition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExhibitionRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}:
    get:
      tags: [Exhibitions]
      summary: 获取展厅详情
      operationId: getExhibitionDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 展厅详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags: [Exhibitions]
      summary: 更新展厅基础信息
      operationId: updateExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExhibitionRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}/members:
    get:
      tags: [Exhibitions]
      summary: 获取协作者列表
      operationId: getExhibitionMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 协作者列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExhibitionMember'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Exhibitions]
      summary: 邀请协作者
      operationId: addExhibitionMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddExhibitionMembersRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}/versions:
    get:
      tags: [Exhibitions]
      summary: 获取版本历史
      operationId: getExhibitionVersions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 版本历史列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExhibitionVersion'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Exhibitions]
      summary: 保存展厅版本
      operationId: saveExhibitionVersion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveExhibitionVersionRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionVersion'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}/viewer:
    get:
      tags: [Exhibitions]
      summary: 获取观众预览数据
      operationId: getExhibitionViewer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 观众模式数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionViewerData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}/publish:
    post:
      tags: [Exhibitions]
      summary: 发布展厅
      operationId: publishExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishExhibitionRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /exhibitions/{exhibitionId}/digital-human:
    get:
      tags: [DigitalHuman]
      summary: 获取展厅数字人详情
      operationId: getExhibitionDigitalHuman
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 数字人详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DigitalHuman'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags: [DigitalHuman]
      summary: 创建或更新数字人
      operationId: upsertExhibitionDigitalHuman
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertDigitalHumanRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DigitalHuman'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /museum/resources:
    get:
      tags: [Museum]
      summary: 获取博物馆资源列表
      operationId: getMuseumResources
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: providerCode
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: keyword
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 文博资源分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedMuseumResource'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Museum]
      summary: 同步博物馆资源
      operationId: syncMuseumResources
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncMuseumResourcesRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /museum/resources/{resourceId}:
    get:
      tags: [Museum]
      summary: 获取博物馆资源详情
      operationId: getMuseumResourceDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 文博资源详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MuseumResource'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /community/exhibitions/{exhibitionId}:
    get:
      tags: [Community]
      summary: 获取社区展厅详情
      operationId: getCommunityExhibitionDetail
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 社区展厅详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionViewerData'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ExhibitionId:
      in: path
      name: exhibitionId
      required: true
      schema:
        type: integer
        format: int64
...
    ResourceId:
      in: path
      name: resourceId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    SuccessResponse:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 0
            message: ok
            data: {}
            requestId: 4f9b6e24e2ef4f4f9d6c
    ErrorResponse:
      description: 业务异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 40100
                message: 未登录或 token 失效
                data: null
                requestId: 4f9b6e24e2ef4f4f9d6c
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: 参数错误
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    UserRole:
      type: string
      enum: [student, teacher, parent, admin, visitor]
...
    ExhibitionStatus:
      type: string
      enum: [draft, published, archived]
    ExhibitionVisibility:
      type: string
      enum: [private, class, public]
...
    SimpleUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        name:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
...
    ExhibitionStats:
      type: object
      properties:
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        commentCount:
          type: integer
    ExhibitionSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        summary:
          type: string
        status:
          $ref: '#/components/schemas/ExhibitionStatus'
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
        groupName:
          type: string
        ownerId:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/SimpleUser'
        latestVersionNo:
          type: integer
        publishedVersionNo:
          type: integer
        stats:
          $ref: '#/components/schemas/ExhibitionStats'
        tags:
          type: array
          items:
            type: string
    ExhibitionDetail:
      allOf:
        - $ref: '#/components/schemas/ExhibitionSummary'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            collaborators:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitionMember'
    ExhibitionMember:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
        joinedAt:
          type: string
          format: date-time
    CreateExhibitionRequest:
      type: object
      required: [title]
      properties:
        taskId:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        coverUrl:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
        groupName:
          type: string
    UpdateExhibitionRequest:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        coverUrl:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
    AddExhibitionMembersRequest:
      type: object
      required: [memberUserIds, role]
      properties:
        memberUserIds:
          type: array
          items:
            type: integer
            format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
    CanvasConfig:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        background:
          type: string
        zoom:
          type: number
          format: double
    ExhibitionElement:
      type: object
      properties:
        componentType:
          type: string
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        props:
          type: object
          additionalProperties: true
    ExhibitionRenderData:
      type: object
      properties:
        canvasConfig:
          $ref: '#/components/schemas/CanvasConfig'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionElement'
    SaveExhibitionVersionRequest:
      type: object
      required: [saveType, canvasConfig, versionData]
      properties:
        saveType:
          type: string
          enum: [manual, autosave, publish]
        versionNote:
          type: string
        canvasConfig:
          $ref: '#/components/schemas/CanvasConfig'
        versionData:
          $ref: '#/components/schemas/ExhibitionRenderData'
    ExhibitionVersion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        versionNo:
          type: integer
        saveType:
          type: string
        versionNote:
          type: string
        canvasConfig:
          $ref: '#/components/schemas/CanvasConfig'
        elementCount:
          type: integer
        versionData:
          $ref: '#/components/schemas/ExhibitionRenderData'
        createdBy:
          $ref: '#/components/schemas/SimpleUser'
        createdAt:
          type: string
          format: date-time
    PublishExhibitionRequest:
      type: object
      required: [versionNo, visibility]
      properties:
        versionNo:
          type: integer
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
...
    StoryTimelineItem:
      type: object
      properties:
        anchorCode:
          type: string
        startSecond:
          type: integer
        endSecond:
          type: integer
        content:
          type: string
    DigitalHumanEquipment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        slotCode:
          type: string
        anchorCode:
          type: string
        displayOrder:
          type: integer
        resourceId:
          type: integer
          format: int64
        resourceTitle:
          type: string
        museumName:
          type: string
        resourceSnapshot:
          type: object
          additionalProperties: true
    DigitalHuman:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exhibitionId:
          type: integer
          format: int64
        name:
          type: string
        avatar2dUrl:
          type: string
          format: uri
        model3dUrl:
          type: string
          format: uri
        persona:
          type: string
        voiceType:
          type: string
        storyScript:
          type: string
        storyTimeline:
          type: array
          items:
            $ref: '#/components/schemas/StoryTimelineItem'
        equippedItems:
          type: array
          items:
            $ref: '#/components/schemas/DigitalHumanEquipment'
    UpsertDigitalHumanRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        avatar2dUrl:
          type: string
          format: uri
        model3dUrl:
          type: string
          format: uri
        persona:
          type: string
        voiceType:
          type: string
        storyScript:
          type: string
        storyTimeline:
          type: array
          items:
            $ref: '#/components/schemas/StoryTimelineItem'
...
    MuseumResource:
      type: object
      properties:
        id:
          type: integer
          format: int64
        providerCode:
          type: string
        title:
          type: string
        category:
          type: string
        museumName:
          type: string
        coverUrl:
          type: string
          format: uri
        detailUrl:
          type: string
          format: uri
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
    SyncMuseumResourcesRequest:
      type: object
      required: [providerCode]
      properties:
        providerCode:
          type: string
        category:
          type: string
        keyword:
          type: string
...
    SubmissionReview:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reviewer:
          $ref: '#/components/schemas/SimpleUser'
        score:
          type: number
          format: double
        commentText:
          type: string
        commentAudioUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exhibitionId:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/SimpleUser'
        parentCommentId:
          type: integer
          format: int64
          nullable: true
        content:
          type: string
        mentionUsers:
          type: array
          items:
            $ref: '#/components/schemas/SimpleUser'
        createdAt:
          type: string
          format: date-time
...
    ExhibitionViewerData:
      type: object
      properties:
        exhibition:
          $ref: '#/components/schemas/ExhibitionDetail'
        renderData:
          $ref: '#/components/schemas/ExhibitionRenderData'
        digitalHuman:
          $ref: '#/components/schemas/DigitalHuman'
        teacherReviews:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionReview'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    PagedExhibitionSummary:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionSummary'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    PagedMuseumResource:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/MuseumResource'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...